// src/app/admin/events/models/operation-type.ts
export enum OperationType {
  CustomerRegistered = 'CustomerRegistered',
  Minted = 'Minted',
  Sent = 'Sent',
  Burned = 'Burned',
  BridgeAdded = 'BridgeAdded',
  LinkerAdded = 'LinkerAdded',
  ManagerAdded = 'ManagerAdded',
  MinterAdded = 'MinterAdded',
  OwnershipTransferred = 'OwnershipTransferred',
  RegistrarAdded = 'RegistrarAdded',
  Transfer = 'Transfer',
  TransferAccepted = 'TransferAccepted',
  TransferReceived = 'TransferReceived',
  TransferRejected = 'TransferRejected',
  StakeDecreased = 'StakeDecreased',
  StakeIncreased = 'StakeIncreased',
  FeeCollected = 'FeeCollected',
  TransferredFromPublicNetwork = 'TransferredFromPublicNetwork',
  TransferredToPublicNetwork = 'TransferredToPublicNetwork',
  TransferToPublicNetworkFeeSet = 'TransferToPublicNetworkFeeSet',
  PublicAccountLinked = 'PublicAccountLinked',
  PublicAccountUnlinked = 'PublicAccountUnlinked',
  RoleAdded = 'RoleAdded',
  SeizeFrom = 'SeizeFrom',
  TreasuryAccountSet = 'TreasuryAccountSet'
}